2-triage-api/requirements.txt
flask
gunicorn
numpy
scikit-learn
google-cloud-storage
flask-cors
requests
